requirements.txt
Django==1.11.20
django-bootstrap3==9.0.0
django-dnt==0.2.0
django-filter==1.0.4
django-fontawesome==0.3.1
django-postgres-copy==0.2.2
django-registration==2.2
djangorestframework==3.6.3
jsonschema==2.6.0
numpy==1.22.2
psycopg2-binary==2.8.6
raven==6.10.0
uWSGI==2.0.18

